J'ai fait une erreur de type
int au lieu de double dans la fonction ajour dans la pile

+++ Code/pile.hpp
@@ -11,7 +11,7 @@ public:
     ~Pile();
 
     bool verifier_pile_vide();
-    void Ajouter_a_la_pile (stack<int> st, int element);
+    void Ajouter_a_la_pile (stack<int> st, double element);
     double lire_premier_elements_pile (stack<double> pile_locale);
     //double lire_deuxiemme_elements_pile (stack<double> pile_locale);
     void supprimer_premier_elements_pile (stack<double> pile_locale);
